refactor: create directory for screens' viewmodels
